Class refactoring. New CameraScanner and ReceiptDownloader classes
